+++ cpp/day_8/1_add_comand_elem.cpp
@@ -3,14 +3,15 @@
 using namespace std;
 int main(int argc, char * argv[])
 {
-	int sum = 0;
+	int sum1 = 0,sum2 = 0;
 	for(int i = 0; i < argc;i++)
 	{
 		if(i != 0)
-		{
-			sum = sum + myAtoi(argv[i]);
+		{ 
+			sum1 += atoi(argv[i]);
+			sum2 += myAtoi(argv[i]);
 		}
 	} 
-	cout << sum << endl;
+	cout << "Original atoi() : " << sum1 << "\n" << "My atoi() : " << sum2 << endl;
 	return 0;
 }
